+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/TabGroup` to read out contents without button `role`
+
 ## [1.0.1] - 2020-07-20
 
 ### Fixed